Migrate BetterEventEditor to EventEditor
