+++ admin-portal/src/app/_components/EventEditor.tsx
@@ -1,110 +1,98 @@
-import React, { useState, useRef, useEffect } from 'react';
-import { EventTags } from "../_interfaces/EventInterfaces";
-import { RootState } from '../_interfaces/AuthInterfaces';
-import QRCode from 'react-qr-code';
-import '../_styles/EventEditor.css';
-import { pen, logo } from '../../../public/EventEditor/EventEditor-index'
-import axios, { AxiosResponse } from 'axios';
-import ActivityDropdown from './ActivityDropdown';
+"use client";
+
+import axios, { AxiosResponse } from "axios";
+import Image from 'next/image';
+import Link from 'next/link';
 import { useRouter } from 'next/navigation';
-import { RequestType } from '../_interfaces/RequestInterfaces';
-import useApiAuth from '../_hooks/useApiAuth';
+import React, { useEffect, useRef, useState } from 'react';
+import { useSelector } from 'react-redux';
+
+import { add_photo, calendar, clock, down_arrow, draft, hero, person, right_arrow } from '../../../public/EventEditor/EventEditor-index';
+import { RootState } from '../_interfaces/AuthInterfaces';
+import { EventTags, LocationProps } from "../_interfaces/EventInterfaces";
+import '../_styles/EventEditor.css';
+
+import ActivityDropdown from './ActivityDropdown';
+import TagModal from "./TagModal";
 
 interface EventEditorProps {
-    orgName: string;
-    changeState: React.Dispatch<React.SetStateAction<boolean>>;
     eventId: string;
-    justCreated: boolean;
 }
 
-interface LocationProps {
-    display_name: string;
-    lat: string;
-    lon: string;
-}
-
-// TODO: Update the organization profile after creating this event to showcase.
-export default function EventEditor({ orgName, changeState, eventId, justCreated }: EventEditorProps) {
-    const [updatedName, setUpdatedName] = useState<string>("Your Event Name");
-    const [updatedDate, setUpdatedDate] = useState<Date>(new Date());
-    const [updatedDescription, setUpdatedDescription] = useState<string>("Your Event Description");
-    const [updatedTags, setUpdatedTags] = useState<boolean[]>(Array(EventTags.length).fill(false));
-    const [currLatitude, setLatitude] = useState<number>(0);
-    const [currLongitude, setLongitude] = useState<number>(0);
+export default function EventEditor({ eventId }: EventEditorProps): React.ReactElement {
+    const router = useRouter();
+    const [eventTitle, setEventTitle] = useState<string>("");
+    const [eventDescription, setEventDescription] = useState<string>("");
+    const [eventAttendeesCount, setEventAttendeesCount] = useState<number>(0);
+    const [latitude, setLatitude] = useState<number>(0);
+    const [longitude, setLongitude] = useState<number>(0);
     const [options, setOptions] = useState<LocationProps[]>([]);
-    const address = useRef<string>("");
+    const [address, setAddress] = useState<string>("");
+    const [tags, setTags] = useState<boolean[]>(Array(EventTags.length).fill(false));
+    const selectedTagColors = ["tag-selected-red", "tag-selected-salmon", "tag-selected-yellow"];
     const [startTime, setStartTime] = useState('12:00');
     const [endTime, setEndTime] = useState('12:01');
+    const [eventDate, setEventDate] = useState<Date>(new Date());
     const [selectedTimeZone, setSelectedTimeZone] = useState('PT');
     const [isDraft, setIsDraft] = useState(true);
-    // A state variable I used to debug Nominatim Issues. Can be useful for notifications
-    const [submissionStatus, setSubmissionStatus] = useState<string>("");
-
-    // The CSS State Variables
-    const [isEditingName, setIsEditingName] = useState<boolean>(false);
-    const [isEditingDescription, setIsEditingDescription] = useState<boolean>(false);
-    // This timer will start when the user stops typing and reset once the user starts typing again
+    const [image, setImage] = useState<string | null>(null);
     const [timeoutID, setTimeoutID] = useState<NodeJS.Timeout>();
-    const [org, sendRequest] = useApiAuth();
 
-    const timeZones = [
-        { label: "Pacific Time (PT) America/Los Angeles", value: "PT" },
-        { label: "Mountain Time (MT) America/Denver", value: "MT" },
-        { label: "Central Time (CT) America/Chicago", value: "CT" },
-        { label: "Eastern Time (ET) America/New York", value: "ET" },
-    ];
-    const router = useRouter();
+    const [isEditingEventTitle, setIsEditingEventTitle] = useState<boolean>(false);
+    const [isEditingDescription, setIsEditingDescription] = useState<boolean>(false);
+    const [isEditingLocation, setIsEditingLocation] = useState<boolean>(false);
+    const [isDisplayingTagModal, setIsDisplayingTagModal] = useState<boolean>(false);
+    const [isShowingDatePicker, setIsShowingDatePicker] = useState<boolean>(false);
+    const [isShowingStartTime, setIsShowingStartTime] = useState<boolean>(false);
+    const [isShowingEndTime, setIsShowingEndTime] = useState<boolean>(false);
+
+    const org = useSelector((state: RootState) => { return { orgId: state.auth.orgId, authToken: state.auth.authToken, refreshToken: state.auth.refreshToken } })
+
+    const dateInputRef = useRef<HTMLInputElement>(null);
+    const startTimeRef = useRef<HTMLInputElement>(null);
+    const endTimeRef = useRef<HTMLInputElement>(null);
 
     useEffect(() => {
         (async () => {
-            // Parse the event's JSON from db
-            // NOTE: This pings the database twice for some reason, doesnt affect correction though, so whatevs
-            // NOTE: We can ignore the draft list as we can generate a new one
-
             const eventData = await getEventById();
-            setUpdatedName(eventData.name ?? "");
-            setUpdatedDate(new Date(eventData.date))
-            // We dont use Duration?
-            setIsDraft(eventData.draft)
-            setUpdatedDescription(eventData.description)
-            setStartTime(eventData.startTime)
-            setEndTime(eventData.endTime)
-            if (eventData.location?.coordinates?.length === 2) {
-                setLatitude(eventData.location.coordinates[0]);
-                setLongitude(eventData.location.coordinates[1]);
-            } else {
-                console.warn('Missing or malformed coordinates in eventData:', eventData.location);
-                setLatitude(0); // or some default/fallback
-                setLongitude(0);
-            }
-            setUpdatedTags(eventData.tags)
-            // TODO: No registeredUsers, activity, image for now ... .. . ...
+            setEventTitle(eventData.name);
+            setIsDraft(eventData.draft);
+            setEventDescription(eventData.description);
+            setEventDate(new Date(eventData.date))
+            setStartTime(eventData.startTime);
+            setEndTime(eventData.endTime);
+            setLatitude(eventData.location.coordinates[0]);
+            setLongitude(eventData.location.coordinates[1]);
+            setTags(eventData.tags);
+            setImage(eventData.imageURL);
+            // TODO: Fix userCount PATCH
+            setEventAttendeesCount(eventData.userCount ?? 0);
         })();
+    }, []);
 
-        if (isEditingDescription) {
-            document.getElementById("description")?.focus();
+    const getEventById = async () => {
+        try {
+            const response: AxiosResponse = await axios.get(`http://${process.env.IP_ADDRESS}:${process.env.PORT}/events/${eventId}`, {
+                headers: {
+                    "Content-Type": "application/json",
+                    "Authorization": `Bearer ${org.authToken}`
+                }
+            });
+            const { data } = response.data;
+            return data;
+        } catch (err) {
+            console.log(err);
+            return err;
         }
-    }, [])
+    };
 
-    // Checks if all inputtables are non-empty
     const notEmpty = () => {
-        return ((updatedName != "Your Event Name") &&
-            (updatedDescription != "Your Event Description") &&
-            ((updatedTags.length > 0)) &&
-            ((currLatitude != 0) && (currLongitude != 0)))
-    }
-
-    const getDayOfWeek = (d: Date) => {
-        const days = ["Sun", "Mon", "Tues", "Wed", "Thurs", "Fri", "Sat"];
-        return days[d.getDay()];
-    }
-
-    const getMonth = (d: Date) => {
-        const months = [
-            "January", "February", "March", "April", "May", "June",
-            "July", "August", "September", "October", "November", "December"
-        ];
-        return months[d.getMonth()];
+        return ((eventTitle.length > 0) &&
+            (eventDescription.length > 0) &&
+            ((tags.length > 0)) &&
+            ((latitude != 0) && (longitude != 0)) &&
+            (eventAttendeesCount > 0))
+        // TODO: Add check for image upload
     }
 
     // When creating draft, create array of flags of "TODOs" for a given event
@@ -113,42 +101,22 @@ export default function EventEditor({ orgName, changeState, eventId, justCreated
         const currDraftList = [];
 
         // Event Name: 0
-        currDraftList.push(updatedName !== "Your Event Name" && updatedName !== "");
+        currDraftList.push(eventTitle !== "");
 
         // Start time, end time, time zone have default vals (they aren't in draft list)
 
         // Description: 1
-        currDraftList.push(updatedDescription !== "Your Event Description" && updatedDescription !== "");
+        currDraftList.push(eventDescription !== "");
 
         // Location: 2
-        currDraftList.push(((currLatitude != 0) && (currLongitude != 0)));
+        currDraftList.push(((latitude != 0) && (longitude != 0)));
 
         // Tags: 3
-        currDraftList.push((updatedTags.length != 0));
+        currDraftList.push((tags.length != 0));
 
         // TODO: When implemented, add flag for event photo
 
         return currDraftList
-    }
-
-    // Date object store dates under the assumption of array accesses, which means last days of the months overflow
-    // This function fixes that, aka: March 32 --> April 1, and so on
-    const parseDate = (d: Date) => {
-        const daysInMonth = [31, 28, 31, 30, 31, 30, 31, 31, 30, 31, 30, 31];
-        const leapYearDaysInMonth = [31, 29, 31, 30, 31, 30, 31, 31, 30, 31, 30, 31];
-        const month = d.getMonth();
-        const year = d.getFullYear();
-        let date = d.getDate();
-
-        // Leap Year
-        if (year % 4 == 0) {
-            date = ((date) >= leapYearDaysInMonth[month]) ? 1 : date + 1;
-        }
-        else {
-            date = ((date) >= daysInMonth[month]) ? 1 : date + 1;
-        }
-
-        return new Date(year, month, date);;
     }
 
     // TODO: maybe refresh to populate the event into org upon successful patch?
@@ -158,86 +126,48 @@ export default function EventEditor({ orgName, changeState, eventId, justCreated
         try {
             // TODO: Convert to GMT, figure out how things are stored
             if ((notEmpty() && !currIsDraft) || currIsDraft) {
-                const selectedTags = updatedTags
+                const selectedTags = tags
                     .map((isSelected, index) => isSelected ? EventTags[index] : null)
                     .filter(tag => tag !== null);
 
                 const uploadDraftList = generateDraftList()
 
-                const body = {
-                    name: updatedName,
-                    date: updatedDate,
-                    duration: 0, // Hardcoded for now
-                    draft: currIsDraft,
-                    draftList: uploadDraftList,
-                    description: updatedDescription,
-                    startTime: startTime,
-                    endTime: endTime,
-                    location: {
-                        type: "Point",
-                        coordinates: [currLongitude, currLatitude]
+                const response: AxiosResponse = await axios.patch(
+                    `http://${process.env.IP_ADDRESS}:${process.env.PORT}/events/${eventId}`,
+                    {
+                        name: eventTitle,
+                        date: eventDate,
+                        duration: 0, // Hardcoded for now
+                        draft: currIsDraft,
+                        draftList: uploadDraftList,
+                        description: eventDescription,
+                        startTime: startTime,
+                        endTime: endTime,
+                        location: {
+                            type: "Point",
+                            coordinates: [longitude, latitude]
+                        },
+                        organizerID: org.orgId,
+                        tags: selectedTags,
+                        registeredUsers: [], // Hardcoded for now
+                        activity: [], // Hardcoded for now
+                        imageURL: image,
+                        userCount: eventAttendeesCount,
                     },
-                    organizerID: org.orgId,
-                    tags: selectedTags,
-                    registeredUsers: [], // Hardcoded for now
-                    activity: [], // Hardcoded for now
-                    image: "placeholder" // Hardcoded for now
-                };
-                const endpoint = `events/${eventId}`;
-                const requestType = RequestType.PATCH;
-                await sendRequest({ requestType, endpoint, body });
-                changeState(false);
+                    {
+                        headers: {
+                            "Content-Type": "application/json",
+                            "Authorization": `Bearer ${org.authToken}`
+                        }
+                    }
+                );
+                router.push('/events');
             }
             else {
-                let errs = "";
-                if (updatedName == "Your Event Name") {
-                    errs = errs + "Name ";
-                }
-                if (updatedDescription == "Your Event Description") {
-                    errs = errs + "Description ";
-                }
-                if (updatedTags.length == 0) {
-                    errs = errs + "Tags ";
-                }
-                if ((currLatitude == 0) && (currLongitude == 0)) {
-                    errs = errs + "Address "
-                }
-                setSubmissionStatus(`Error: Empty Args: ${errs}`);
+                console.log("Upload Failure: Empty Arguments")
             }
         } catch (err) {
             console.log(err);
-            setSubmissionStatus(`Failure: ${err}`);
-        }
-    }
-
-    const getEventById = async () => {
-        try {
-            const requestType: RequestType = RequestType.GET;
-            const body = {};
-            const endpoint = `events/${eventId}`;
-            const data = await sendRequest({ requestType, body, endpoint })
-            return data;
-        } catch (err) {
-            console.log(err);
-            return err;
-        }
-    };
-
-    // Use HTML GeoLocation API to get the user's location (shall their web browser permit)
-    // Updates state variables latitude, longitude if it works
-    const getUserLocation = () => {
-        if (navigator.geolocation) {
-            navigator.geolocation.getCurrentPosition(
-                (position) => {
-                    const { latitude, longitude } = position.coords;
-                    const query = `${latitude}, ${longitude}`;
-                    getLocationJSON(query);
-                    setOptions([]);
-                },
-                (err) => {
-                    console.error("Error: Could not retrieve location", err);
-                }
-            );
         }
     }
 
@@ -277,7 +207,6 @@ export default function EventEditor({ orgName, changeState, eventId, justCreated
             }
             // No address returned
             else {
-                setSubmissionStatus("Error: Invalid Address");
                 setLatitude(0);
                 setLongitude(0);
             }
@@ -286,249 +215,339 @@ export default function EventEditor({ orgName, changeState, eventId, justCreated
         }
     }
 
-    return (
-        <div className="box">
-            {/* The left side */}
-            <div className="left">
-                {/* Image selection */}
-                <div className="blankImage">
-                    <div className="imagebox">
-                        <img className="penlogo" src={pen.src} />
-                    </div>
+    const handleImageChange = async (e: React.ChangeEvent<HTMLInputElement>) => {
+        if (e.target.files && e.target.files[0]) {
+            const file = e.target.files[0];
+            const reader = new FileReader();
+            reader.onloadend = () => {
+                setImage(reader.result as string);
+            };
+            reader.readAsDataURL(file);
+            // Upload the image
+            const imageUrl = await uploadOrgImage(file);
+            console.log("Uploaded image URL:", imageUrl);
+            setImage(imageUrl);
+        }
+    };
 
-                    <div className="leftBottom">
-                        <div className="customizeText">
-                            Customize your Event
+    const uploadOrgImage = async (file: File) => {
+        console.log("Image upload token:", org.authToken);
+        const formData = new FormData();
+        formData.append("image", file);
+        const response = await axios.post(
+            `http://${process.env.IP_ADDRESS}:${process.env.PORT}/orgs/${org.orgId}/addImage`,
+            formData,
+            {
+                headers: {
+                    Authorization: `Bearer ${org.authToken}`,
+                    "Content-Type": "multipart/form-data"
+                },
+            }
+        );
+        console.log("Image upload response:", response.data);
+        return response.data.imageUrl;
+    };
+
+    function formatToAMPM(d: Date): string {
+        const date = new Date(d);
+
+        let hours = date.getHours();
+        const minutes = date.getMinutes();
+
+        const ampm = hours >= 12 ? 'PM' : 'AM';
+        hours = hours % 12;
+        hours = hours ? hours : 12; // the hour '0' should be '12'
+        const minutesStr = minutes < 10 ? '0' + minutes : minutes;
+
+        return `${hours}:${minutesStr} ${ampm}`;
+    }
+
+    const formatDate = (date: Date | null): string => {
+        if (!date) return '--/--/----';
+
+        const month = (date.getMonth() + 1).toString().padStart(2, '0');
+        const day = date.getDate().toString().padStart(2, '0');
+        const year = date.getFullYear();
+
+        return `${month}/${day}/${year}`;
+    }
+
+    const formatTimeToAMPM = (time: string): string => {
+        const [hours, minutes] = time.split(':');
+        const hour = parseInt(hours);
+        const ampm = hour >= 12 ? 'PM' : 'AM';
+        const displayHour = hour % 12 || 12;
+        return `${displayHour}:${minutes} ${ampm}`;
+    };
+
+    return (
+        <div>
+            {/* Events header */}
+            <div className="event-header">
+                <div className="event-header-info">
+                    <div className="event-directory-path">
+                        <div>Home</div>
+                        <div className="right-arrow"><Image src={right_arrow} alt="Right Arrow Icon" width={8} height={13} /></div>
+                        <div>Events</div>
+                        <div className="right-arrow"><Image src={right_arrow} alt="Right Arrow Icon" width={8} height={13} /></div>
+                        <div className="event-directory-path-eventname">{eventTitle.length == 0 ? "Untitled Event" : eventTitle}</div>
+                    </div>
+                </div>
+                <div className="event-header-events-title">EVENTS</div>
+                <div className="event-header-events-subtitle">
+                    View current, published, past, and event drafts.
+                </div>
+            </div>
+            {/* Editor Body */}
+            <div className="editor-body-parent">
+                <div className="draft-save-publish-indicators">
+                    <div className="draft-indicator">
+                        <div className="draft-indicator-logo"><Image src={draft} alt="Draft Icon" width={24} height={24} /></div>
+                        <div className="draft-indicator-text">DRAFT</div>
+                    </div>
+                    <div className="cancel-save-publish-parent">
+                        <Link className="cancel-button" href="/events">CANCEL</Link>
+                        <div className="save-button" onClick={() => { handlePatch(true) }}>SAVE</div>
+                        <div className="publish-button" onClick={() => { handlePatch(false) }}>PUBLISH</div>
+                    </div>
+                </div>
+
+                <div className="event-editor-interface">
+                    <div className="image-editor-and-tags">
+                        {/* This setup creates a hidden input that's triggered by clicking the label */}
+                        <input
+                            type="file"
+                            accept="image/*"
+                            onChange={handleImageChange}
+                            className="hidden"
+                            id="imageUpload"  // This ID connects the input to the label
+                        />
+                        {
+                            image ?
+                                <label htmlFor="imageUpload" className="image-editor-with-image">
+                                    <div className="add_photo-image-with-image">
+                                        <Image src={image} alt="Add Photo Icon" width={513} height={450} />
+                                    </div>
+                                </label>
+                                :
+                                <label htmlFor="imageUpload" className="image-editor">
+                                    <label htmlFor="imageUpload" className="add-photo-parent">
+                                        <div className="add_photo-image">
+                                            <Image src={add_photo} alt="Add Photo Icon" width={60} height={60} />
+                                        </div>
+                                        <div className="add_photo-subtitle">
+                                            Upload Images Here
+                                        </div>
+                                    </label>
+                                </label>
+                        }
+                        <div className="tags-title">
+                            SELECT KEYWORDS
                         </div>
-                        <div className="customizeButtonFormat">
-                            <button className="customizeButton">
-                                Add Poll
-                            </button>
-                            <button className="customizeButton">
-                                Add Rewards
-                            </button>
-                            <div style={{ display: 'flex' }}>
-                                <QRCode value={eventId} />
+                        <div className="tags-parent">
+                            <div className="tags-container">
+                                {
+                                    isDisplayingTagModal ?
+                                        (<TagModal tags={tags} setTags={setTags} setIsDisplayingTagModal={setIsDisplayingTagModal} />)
+                                        :
+                                        EventTags.map((_, index) => {
+                                            return (
+                                                tags[index] ?
+                                                    <div
+                                                        className={selectedTagColors[(index) % 3]}
+                                                        key={index}>
+                                                        {EventTags[index]}
+                                                    </div>
+                                                    :
+                                                    <React.Fragment key={index}></React.Fragment>
+                                            )
+                                        })
+                                }
+                            </div>
+                            <div className="add-tag-button" onClick={() => { setIsDisplayingTagModal(true) }}>
+                                + ADD TAG
                             </div>
                         </div>
                         <ActivityDropdown eventId={eventId} isDraft={true} />
                     </div>
-                </div>
-            </div>
-            {/* The right side */}
-            <div className="right">
-
-                {/* Cancel and Publish Buttons */}
-                <div className="goToTheRight">
-                    <button className="saveButton" onClick={() => {
-                        handlePatch(true);
-                        router.push('/events');
-                    }}>
-                        SAVE
-                    </button>
-                    <button className="bigPillButton" onClick={async () => {
-                        changeState(false)
-                        // If event just created, delete the event
-                        // NOTE: This is not async, might cause problems later
-                        if (justCreated) {
-                            const requestType: RequestType = RequestType.DELETE;
-                            const body = {};
-                            const endpoint = `events/${eventId}`;
-                            await sendRequest({ requestType, body, endpoint });
-                        }
-                        router.push('/events');
-                    }}>
-                        CANCEL
-                    </button>
-                    <button className="bigPillButton" onClick={() => {
-                        handlePatch(false);
-                        router.push('/events');
-                    }}>
-                        PUBLISH
-                    </button>
-                </div>
-
-                {/* Event Name */}
-                <div className="eventBox" onClick={() => setIsEditingName(true)}>
-                    {
-                        // If is editing, then put in a textbox. If not, then its plain text
-                        isEditingName ?
-                            <textarea
-                                onKeyDown={(e) => {
-                                    if (e.key === "Enter") {
-                                        setIsEditingName(false)
-                                    }
-                                }}
-                                onBlur={() => { setIsEditingName(false); }}
-                                className="titleinputbox" name="name" placeholder="Name" value={updatedName} onChange={(event) => { setUpdatedName(event.target.value) }}
-                            />
-                            :
-                            <div className="titleinputbox">{updatedName}</div>
-                    }
-                </div>
-
-                {/* Organization Info */}
-                <div className="orgInfoBox">
-                    <div className="orgLogo">
-                        <img src={logo.src} />
-                    </div>
-                    <div className="flexIt">
-                        <h4>
-                            Hosted by <u style={{ color: '#696969' }}>{orgName}</u>:
-                        </h4>
-                    </div>
-                    <button className="editButton">
-                        EDIT
-                    </button>
-                </div>
-
-                {/* Date, Time, Timezone */}
-                <div className="dateRow">
-                    <div className="dateTimeZoneBox">
-                        {/* Date */}
-                        <div
-                            className="date"
-                            onClick={() => (document.getElementById('hiddenDateInput') as HTMLInputElement).showPicker()}
-                        >
-                            <div>{getDayOfWeek(updatedDate)}</div>
-                            <div>{getMonth(updatedDate)}</div>
-                            <div>{updatedDate.getDate()}</div>
+                    <div className="event-detail-fields">
+                        <div className="event-title-and-org">
+                            <div className="event-name-title">EVENT NAME</div>
+                            <div onClick={() => setIsEditingEventTitle(true)}>
+                                {
+                                    isEditingEventTitle ?
+                                        <textarea
+                                            onKeyDown={(e) => {
+                                                if (e.key === "Enter") { setIsEditingEventTitle(false) }
+                                            }}
+                                            onBlur={() => { setIsEditingEventTitle(false); }}
+                                            className="event-title-editor-card" name="eventTitle" placeholder="TITLE" value={eventTitle} onChange={(event) => { setEventTitle(event.target.value) }}
+                                        />
+                                        :
+                                        <div className="event-title-editor-card">{eventTitle.length == 0 ? (<div className="event-title-editor-card-empty">TITLE</div>) : eventTitle}</div>
+                                }
+                            </div>
+                            <div className="organization-attendees-grandparent">
+                                <div className="org-box-parent">
+                                    <div className="org-logo"><Image src={hero} alt="Hero Icon" width={28} height={28} /></div>
+                                    <div className="org-info">
+                                        Hosted by
+                                        <div className="org-name">United Way</div>
+                                    </div>
+                                </div>
+                                <div className="attendees-parent">
+                                    <div className="attendees-image-and-title">
+                                        <div className="attendees-image"><Image src={person} alt="Attendees Icon" width={15} height={17} /></div>
+                                        <div className="attendees-box-title">MAX ATTENDEES</div>
+                                        <div className="date-box-vertical-bar" />
+                                    </div>
+                                    <div className="attendees-counter">
+                                        <div className="attendees-plus-minus" onClick={() => { setEventAttendeesCount(Math.max(eventAttendeesCount - 1, 0)) }}>–</div>
+                                        <div className="attendees-count">{eventAttendeesCount}</div>
+                                        <div className="attendees-plus-minus" onClick={() => { setEventAttendeesCount(eventAttendeesCount + 1) }}>+</div>
+                                    </div>
+                                </div>
+                            </div>
                         </div>
 
-                        {/* Hidden Date Input Interface, only the input modal appears when clicked */}
-                        <input
-                            type="date"
-                            id="hiddenDateInput"
-                            className="hiddenModal flexIt"
-                            value={
-                                updatedDate instanceof Date && !isNaN(updatedDate.getTime())
-                                    ? updatedDate.toISOString().split('T')[0]
-                                    : ''
-                            }
-                            onChange={(event) => { setUpdatedDate(parseDate(new Date((event.target as HTMLInputElement).value))) }}
-                        />
-                    </div>
+                        <div className="date-time-grandparent">
+                            <div className="date-box-parent">
+                                <div className="date-image-and-title">
+                                    <div className="calendar-image"><Image src={calendar} alt="Calendar Icon" width={15} height={17} /></div>
+                                    <div className="date-box-title">DATE</div>
+                                    <div className="date-box-vertical-bar" />
+                                </div>
+                                <div className="date-box-date-and-down"
+                                    onClick={() => {
+                                        dateInputRef.current?.showPicker();
+                                        setIsShowingDatePicker(true);
+                                    }}
+                                    onMouseDown={(e) => {
+                                        e.preventDefault(); // Prevent focus from being lost
+                                    }}>
+                                    <input
+                                        ref={dateInputRef}
+                                        type="date"
+                                        onChange={(e) => setEventDate(new Date(e.target.value))}
+                                        className="date-input"
+                                        value={eventDate.toISOString().split('T')[0]}
+                                        onBlur={() => setIsShowingDatePicker(false)}
+                                    />
+                                    <div className="down_arrow">
+                                        <Image src={down_arrow} alt="Down Arrow Icon" width={15} height={15} />
+                                    </div>
+                                </div>
+                            </div>
+                            <div className="time-box-parent">
+                                <div className="time-image-and-title">
+                                    <div className="clock-image"><Image src={clock} alt="Clock Icon" width={15} height={17} /></div>
+                                    <div className="time-box-title">TIME</div>
+                                    <div className="date-box-vertical-bar" />
+                                </div>
+                                <div className="time-box-time-and-down"
+                                    onClick={(e) => {
+                                        if (e.target === e.currentTarget || (e.target as HTMLElement).classList.contains('time-display')) {
+                                            !isShowingStartTime ? startTimeRef.current?.showPicker() : endTimeRef.current?.showPicker();
+                                            setIsShowingStartTime(!isShowingStartTime);
+                                        }
+                                    }}
+                                    onMouseDown={(e) => e.preventDefault()}>
+                                    <div className="time-display">{`${formatTimeToAMPM(startTime)} - ${formatTimeToAMPM(endTime)}`}</div>
+                                    <input
+                                        ref={startTimeRef}
+                                        type="time"
+                                        onChange={(e) => setStartTime(e.target.value)}
+                                        onBlur={() => setIsShowingStartTime(false)}
+                                        className="time-input"
+                                        value={startTime}
+                                    />
+                                    <input
+                                        ref={endTimeRef}
+                                        type="time"
+                                        onChange={(e) => setEndTime(e.target.value)}
+                                        onBlur={() => setIsShowingEndTime(false)}
+                                        className="time-input"
+                                        value={endTime}
+                                    />
+                                    <div className="down_arrow">
+                                        <Image src={down_arrow} alt="Down Arrow Icon" width={15} height={15} />
+                                    </div>
+                                </div>
+                            </div>
+                        </div>
 
-                    <div className="timebox">
-                        <input
-                            type="time"
-                            className="time"
-                            value={startTime}
-                            onChange={(event) => setStartTime(event.target.value)}
-                        />
-
-                        <div className="time">–</div>
-
-                        <input
-                            type="time"
-                            className="time"
-                            value={endTime}
-                            onChange={(event) => setEndTime(event.target.value)}
-                        />
-
-
-                        <div>•</div>
-
-                        {/* Time Zone */}
-                        {/* For whatever reason, the code breaks if you remove the position: relative in-place design */}
-                        <div className="time" style={{ position: 'relative' }}>
-                            <select
-                                className="timezone"
-                                value={selectedTimeZone}
-                                onChange={(event: React.ChangeEvent<HTMLSelectElement>) => {
-                                    setSelectedTimeZone(event.target.value);
-                                }}
-                            >
-                                {timeZones.map((zone) => (
-                                    <option key={zone.value} value={zone.value}>{zone.label}</option>
-                                ))}
-                            </select>
-                            <div className="clickable">
-                                {selectedTimeZone}
+                        <div className="description-location-attendees">
+                            <div className="description-parent">
+                                <div className="description-title">DESCRIPTION</div>
+                                <div onClick={() => setIsEditingDescription(true)}>
+                                    {
+                                        isEditingDescription ?
+                                            <textarea
+                                                onKeyDown={(e) => {
+                                                    if (e.key === "Enter") { setIsEditingDescription(false) }
+                                                }}
+                                                onBlur={() => { setIsEditingDescription(false); }}
+                                                className="description-body" name="description" placeholder="Description" value={eventDescription} onChange={(event) => { setEventDescription(event.target.value) }}
+                                            />
+                                            :
+                                            <div className="description-body">
+                                                {(eventDescription?.length ?? 0) === 0 ?
+                                                    (<div className="description-body-empty">Description</div>)
+                                                    :
+                                                    eventDescription
+                                                }
+                                            </div>
+                                    }
+                                </div>
+                            </div>
+                            <div className="location-parent">
+                                <div className="location-title">ADDRESS</div>
+                                <div onClick={() => setIsEditingLocation(true)}>
+                                    {
+                                        isEditingLocation ?
+                                            <textarea
+                                                onChange={(e) => {
+                                                    // If an existing timeout exists, kill it (because we're going to set a new one)
+                                                    if (timeoutID) {
+                                                        clearTimeout(timeoutID);
+                                                    }
+                                                    setAddress(e.target.value);
+                                                    const newTimeoutID = setTimeout(() => getLocationJSON(e.target.value), 500);
+                                                    setTimeoutID(newTimeoutID);
+                                                }}
+                                                onKeyDown={(e) => {
+                                                    if (e.key === "Enter") { setIsEditingLocation(false) }
+                                                }}
+                                                onBlur={() => { setIsEditingLocation(false); }}
+                                                className="location" name="location" placeholder="Location" value={address}
+                                            />
+                                            :
+                                            <div className="location">{address.length == 0 ? (<div className="location-empty">Location</div>) : address}</div>
+                                    }
+                                </div>
+                                {/* If multiple results return, this is the modal that pops up */}
+                                <div className="searchOptionsParent">
+                                    <div className="searchOptions">
+                                        {
+                                            options.map((option, index) => (
+                                                <button key={index}
+                                                    onClick={() => {
+                                                        setLatitude(parseFloat(option.lat));
+                                                        setLongitude(parseFloat(option.lon));
+                                                        setAddress(option.display_name);
+                                                        setOptions([]);
+                                                    }}>
+                                                    {option.display_name}
+                                                </button>
+                                            ))
+                                        }
+                                    </div>
+                                </div>
                             </div>
                         </div>
                     </div>
                 </div>
-
-                {/* Description */}
-                <div className="descriptionTitle">
-                    Description:
-                </div>
-
-                <div className="clickable" onClick={() => { setIsEditingDescription(true) }} onBlur={() => setIsEditingDescription(false)}>
-                    {
-                        // If is editing, then put in a textbox. If not, then its plain text
-                        isEditingDescription ?
-                            <textarea id="description" onKeyDown={(e) => {
-                                if (e.key === "Enter") {
-                                    setIsEditingDescription(false)
-                                }
-                            }} className="descriptionInputBox" name="description" placeholder="Description" value={updatedDescription} onChange={(event) => { setUpdatedDescription(event.target.value) }}
-                            />
-                            :
-                            <div className="descriptionInputBox">{updatedDescription}</div>
-                    }
-                </div>
-
-                {/* Location */}
-                <div>
-                    <div className="flexIt">
-                        <textarea className="locationInputbox" onChange={(e) => {
-                            // If an existing timeout exists, kill it (because we're going to set a new one)
-                            if (timeoutID) {
-                                clearTimeout(timeoutID);
-                            }
-                            address.current = e.target.value;
-                            const newTimeoutID = setTimeout(() => getLocationJSON(e.target.value), 500);
-                            setTimeoutID(newTimeoutID);
-                        }} name="location" placeholder="Location" value={address.current}
-                        />
-                    </div>
-                    {/* If multiple results return, this is the modal that pops up */}
-                    <div className="searchOptionsParent">
-                        <div className="searchOptions">
-                            {
-                                options.map((option, index) => (
-                                    <button key={index}
-                                        onClick={() => {
-                                            setLatitude(parseFloat(option.lat));
-                                            setLongitude(parseFloat(option.lon));
-                                            address.current = option.display_name;
-                                            setOptions([]);
-                                        }}>
-                                        {option.display_name}
-                                    </button>
-                                ))
-                            }
-                        </div>
-                    </div>
-                </div>
-
-                <div className="selectKeywordText">
-                    Select Keywords:
-                </div>
-
-                {/* Keyword Buttons */}
-                <div className="tagOptions flexIt">
-                    {
-                        EventTags.map((tagName, index) => {
-                            return (
-                                <button
-                                    className={updatedTags != null && updatedTags[index] ? "tagPillSelected" : "tagPillNotSelected"}
-                                    key={index}
-                                    onClick={() => {
-                                        const newTags = [...updatedTags];
-                                        newTags[index] = !newTags[index];
-                                        setUpdatedTags(newTags);
-                                    }}
-                                >
-                                    {tagName}
-                                </button>
-                            )
-                        })
-                    }
-                </div>
             </div>
         </div>
-    );
+    )
 }
